+++ src/scirpy/ir_dist/_clonotype_neighbors.py
@@ -1,6 +1,6 @@
 import itertools
+import os
 from collections.abc import Mapping, Sequence
-from multiprocessing import cpu_count
 from typing import Literal, Optional, Union
 
 import numpy as np
@@ -28,7 +28,7 @@ class ClonotypeNeighbors:
         match_columns: Union[None, Sequence[str]] = None,
         distance_key: str,
         sequence_key: str,
-        n_jobs: Union[int, None] = None,
+        n_jobs: int = -1,
         chunksize: int = 2000,
     ):
         """Computes pairwise distances between cells with identical
@@ -160,7 +160,9 @@ class ClonotypeNeighbors:
                 )
 
             self.neighbor_finder.add_distance_matrix(
-                "v_gene", sp.identity(len(v_genes), dtype=bool, format="csr"), v_genes  # type: ignore
+                "v_gene",
+                sp.identity(len(v_genes), dtype=bool, format="csr"),
+                v_genes,  # type: ignore
             )
 
         if self.match_columns is not None:
@@ -223,12 +225,12 @@ class ClonotypeNeighbors:
             dist_rows = process_map(
                 self._dist_for_clonotype,
                 range(n_clonotypes),
-                max_workers=self.n_jobs if self.n_jobs is not None else cpu_count(),
+                max_workers=self.n_jobs if self.n_jobs > 0 else len(os.sched_getaffinity(0)),
                 chunksize=2000,
                 tqdm_class=tqdm,
             )
 
-        dist = sp.vstack(dist_rows)
+        dist = sp.vstack(list(dist_rows))
         dist.eliminate_zeros()
         logging.hint("Done computing clonotype x clonotype distances. ", time=start)
         return dist  # type: ignore