+++ src/scirpy/util/__init__.py
@@ -1,3 +1,4 @@
+import contextlib
 import warnings
 from collections.abc import Mapping, Sequence
 from textwrap import dedent
@@ -8,6 +9,7 @@ import numpy as np
 import pandas as pd
 import scipy.sparse
 from anndata import AnnData
+from joblib import Parallel
 from mudata import MuData
 from scanpy import logging
 from scipy.sparse import issparse
@@ -57,13 +59,11 @@ class DataHandler:
 
     @overload
     @staticmethod
-    def default(data: None) -> None:
-        ...
+    def default(data: None) -> None: ...
 
     @overload
     @staticmethod
-    def default(data: "DataHandler.TYPE") -> "DataHandler":
-        ...
+    def default(data: "DataHandler.TYPE") -> "DataHandler": ...
 
     @staticmethod
     def default(data):
@@ -117,12 +117,10 @@ class DataHandler:
             index_chains(self.adata, airr_key=self._airr_key, key_added=self._chain_idx_key)
 
     @overload
-    def get_obs(self, columns: str) -> pd.Series:
-        ...
+    def get_obs(self, columns: str) -> pd.Series: ...
 
     @overload
-    def get_obs(self, columns: Sequence[str]) -> pd.DataFrame:
-        ...
+    def get_obs(self, columns: Sequence[str]) -> pd.DataFrame: ...
 
     def get_obs(self, columns):
         """\
@@ -560,3 +558,18 @@ def _translate_dna_to_protein(dna_seq: str):
         else:
             protein.append("N")
     return "".join(protein)
+
+
+def _parallelize_with_joblib(delayed_objects, *, total=None, **kwargs):
+    """Wrapper around joblib.Parallel that shows a progressbar if the backend supports it.
+
+    Progressbar solution from https://stackoverflow.com/a/76726101/2340703
+    """
+    try:
+        return tqdm(Parallel(return_as="generator", **kwargs)(delayed_objects), total=total)
+    except ValueError:
+        logging.info(
+            "Backend doesn't support return_as='generator'. No progress bar will be shown. "
+            "Consider setting verbosity in joblib.parallel_config"
+        )
+        return Parallel(return_as="list", **kwargs)(delayed_objects)